requirements.txt
edge-tts
asyncio
openai-whisper
langchain
langchain-community